feat: Adiciona função transformObject em Utilitaries
essa função substitui a estrutura de transformar o Objeto do usuário em um objeto mais simples que será lido pela IA. antes estava declarada direta em weekly diet
